UpgradeTrialDialog to show end of trial date
