+++ frontend/src/component/admin/billing/TrialUpsell/TrialUpsell.tsx
@@ -1,6 +1,5 @@
 import { useContext } from 'react';
-import { Alert, Box, Button, Link, styled, Typography } from '@mui/material';
-import OpenInNewIcon from '@mui/icons-material/OpenInNew';
+import { Alert, Box, Button, styled, Typography } from '@mui/material';
 import { formatApiPath } from 'utils/formatPath';
 import { ReactComponent as LenovoLogo } from 'assets/logos/lenovo.svg';
 import AccessContext from 'contexts/AccessContext';
@@ -12,9 +11,7 @@ import { ReactComponent as VisaLogo } from 'assets/logos/visa.svg';
 import { ReactComponent as SamsungLogo } from 'assets/logos/samsung.svg';
 import { ReactComponent as LloydsLogo } from 'assets/logos/lloyds.svg';
 import { trialHasExpired } from 'utils/instanceTrial';
-import { parseISO } from 'date-fns';
-import { formatDateYMD } from 'utils/formatDate';
-import { useLocationSettings } from 'hooks/useLocationSettings';
+import { TrialUpsellDescription } from './TrialUpsellDescription';
 
 const StyledBillingInformation = styled('div')(({ theme }) => ({
     marginTop: theme.spacing(2),
@@ -60,17 +57,10 @@ const StyledLogoList = styled('div')(({ theme }) => ({
     color: theme.palette.text.secondary,
 }));
 
-const formatTrialExpiry = (locale: string, trialExpiry?: string): string => {
-    if (!trialExpiry) return '';
-
-    return formatDateYMD(parseISO(trialExpiry), locale);
-};
-
 export const TrialUpsell = () => {
     const { hasAccess } = useContext(AccessContext);
     const { instancePrices } = useInstancePrices();
     const { instanceStatus } = useInstanceStatus();
-    const { locationSettings } = useLocationSettings();
 
     const onUpgrade = () => {
         window.location.assign(formatApiPath('api/admin/invoices/checkout'));
@@ -78,17 +68,7 @@ export const TrialUpsell = () => {
 
     const isAdmin = hasAccess(ADMIN);
     const seatPrice = instancePrices.payg.seat;
-    const trialExpiry = instanceStatus?.trialExpiry;
     const trialExpired = trialHasExpired(instanceStatus);
-    const locale = locationSettings?.locale;
-
-    const expiryText = trialExpiry
-        ? `on ${formatTrialExpiry(locale, trialExpiry)}`
-        : 'soon';
-
-    const descriptionText = trialExpired
-        ? 'Your free trial has expired and your account is scheduled for deletion. Upgrade to preserve your projects and feature flags.'
-        : `Your trial expires ${expiryText}. Upgrade now to get uninterrupted usage of Unleash.`;
 
     return (
         <>
@@ -97,22 +77,7 @@ export const TrialUpsell = () => {
                     ? 'Upgrade to continue using Unleash'
                     : '14 day free trial'}
             </Typography>
-            <Typography variant='body2'>
-                {descriptionText}{' '}
-                <Link
-                    href='https://www.getunleash.io/pricing'
-                    target='_blank'
-                    rel='noreferrer'
-                    sx={{
-                        display: 'inline-flex',
-                        alignItems: 'center',
-                        gap: 0.5,
-                    }}
-                >
-                    View full pricing
-                    <OpenInNewIcon fontSize='inherit' />
-                </Link>
-            </Typography>
+            <TrialUpsellDescription />
             {isAdmin ? (
                 <StyledBillingInformation>
                     <StyledBillingInformationTextContainer>
