+++ frontend/src/component/common/InstanceStatus/UpgradeTrialDialog.tsx
@@ -11,8 +11,7 @@ import { ReactComponent as VisaLogo } from 'assets/logos/visa.svg';
 import { ReactComponent as SamsungLogo } from 'assets/logos/samsung.svg';
 import { ReactComponent as LloydsLogo } from 'assets/logos/lloyds.svg';
 import { useInstancePrices } from 'hooks/api/getters/useInstancePrices/useInstancePrices';
-import { useInstanceStatus } from 'hooks/api/getters/useInstanceStatus/useInstanceStatus';
-import { trialHasExpired } from 'utils/instanceTrial';
+import { TrialUpsellDescription } from 'component/admin/billing/TrialUpsell/TrialUpsellDescription';
 
 const StyledBillingInformation = styled('div')(({ theme }) => ({
     marginTop: theme.spacing(3),
@@ -55,7 +54,6 @@ export const UpgradeTrialDialog = ({
     dialogOpen,
     setDialogOpen,
 }: IUpgradeTrialDialogProps) => {
-    const { instanceStatus } = useInstanceStatus();
     const { hasAccess } = useContext(AccessContext);
     const { trackEvent } = usePlausibleTracker();
     const { instancePrices } = useInstancePrices();
@@ -96,10 +94,6 @@ export const UpgradeTrialDialog = ({
         setDialogOpen(false);
     };
 
-    const text = trialHasExpired(instanceStatus)
-        ? 'Your free trial has expired and your account is scheduled for deletion. Upgrade to preserve your projects and feature flags.'
-        : 'Upgrade now to unlock all features and keep your projects and feature flags after your trial ends.';
-
     return (
         <Dialogue
             open={dialogOpen}
@@ -110,7 +104,7 @@ export const UpgradeTrialDialog = ({
             title='Upgrade to continue using Unleash'
             maxWidth='sm'
         >
-            {text}
+            <TrialUpsellDescription />
             {isAdmin ? (
                 <StyledBillingInformation>
                     <Typography>
